Transaction SAL complete with all tests passing

+++ SAL/TransactionSAL/TransactionSALImplementation.py
@@ -1,10 +1,14 @@
 import logging
 from typing import List
 
+from DAL.BankAccountDAL.BankAccountDALImplementation import BankAccountDALImplementation
 from DAL.TransactionDAL.TransactionDALImplementation import TransactionDALImplementation
+from Entities.FailedTransaction import FailedTransaction
 from Entities.Transaction import Transaction
 from SAL.TransactionSAL.TransactionSALInterface import TransactionSALInterface
 
+
+account_dao = BankAccountDALImplementation()
 
 class TransactionSALImplementation(TransactionSALInterface):
 
@@ -13,23 +17,77 @@ class TransactionSALImplementation(TransactionSALInterface):
 
     def service_create_transaction(self, transaction: Transaction) -> Transaction:
         logging.info("Beginning SAL method create transaction")
-        logging.info("Finishing SAL method create transaction")
-        return transaction
+        if type(transaction.date_time) != str:
+            logging.info("SAL method create transaction, date and time not a string")
+            raise FailedTransaction("The date and time field must be a string, please try again!")
+        elif len(transaction.date_time) > 26:
+            logging.warning("SAL method create transaction, date and time longer than 26 characters")
+            raise FailedTransaction("The date and time field cannot exceed 26 characters, please try again!")
+        elif len(transaction.date_time) == 0:
+            logging.warning("SAL method create transaction, date and time left empty")
+            raise FailedTransaction("The date and time field cannot be left empty, please try again!")
+        elif type(transaction.transaction_type) != str:
+            logging.warning("SAL method create transaction, transaction type not a string")
+            raise FailedTransaction("The transaction type field must be a string, please try again!")
+        elif len(transaction.transaction_type) > 8:
+            logging.warning("SAL method create transaction, transaction type longer than 8 characters")
+            raise FailedTransaction("The transaction type field cannot exceed 8 characters, please try again!")
+        elif len(transaction.transaction_type) == 0:
+            logging.warning("SAL method create transaction, transaction type left empty")
+            raise FailedTransaction("The transaction type field cannot be left empty, please try again!")
+        elif transaction.transaction_type != ("deposit" or "withdraw" or "transfer" or "expense"):
+            logging.warning("SAL method create transaction, transaction type not deposit, withdraw, transfer, or "
+                            "expense")
+            raise FailedTransaction("The transaction type field must be one of the following: deposit, withdraw, "
+                                    "transfer, expense; please try again!")
+        elif type(transaction.account_id) != int:
+            logging.warning("SAL method create transaction, account ID not an integer")
+            raise FailedTransaction("The account ID field must be an integer, please try again!")
+        elif type(transaction.amount) != float:
+            logging.warning("SAL method create transaction, transaction amount not a float")
+            raise FailedTransaction("The transaction amount field must be a float, please try again!")
+        elif transaction.amount < 0.00:
+            logging.warning("SAL method create transaction, transaction amount negative")
+            raise FailedTransaction("The transaction amount field cannot be negative, please try again!")
+        else:
+            valid_account_check = account_dao.get_account_by_id(transaction.account_id)
+            if valid_account_check is not None:
+                result = self.transaction_dao.create_transaction(transaction)
+                logging.info("Finishing SAL method create transaction")
+                return result
 
     def service_get_transaction_by_id(self, transaction_id: int) -> Transaction:
         logging.info("Beginning SAL method get transaction by ID")
-        result = self.transaction_dao.get_transaction_by_id(transaction_id)
-        logging.info("Finishing SAL method get transaction by ID")
-        return result
+        if type(transaction_id) != int:
+            logging.warning("SAL method get transaction by ID, transaction ID not an integer")
+            raise FailedTransaction("The transaction ID must be an integer, please try again!")
+        else:
+            result = self.transaction_dao.get_transaction_by_id(transaction_id)
+            logging.info("Finishing SAL method get transaction by ID")
+            return result
 
     def service_get_all_transactions(self, account_id: int) -> List[Transaction]:
         logging.info("Beginning SAL method get all transactions")
-        transaction_list = []
-        logging.info("Finishing SAL method get all transactions")
-        return transaction_list
+        if type(account_id) != int:
+            logging.warning("SAL method get all transactions, account ID not an integer")
+            raise FailedTransaction("The account ID field must be an integer, please try again!")
+        else:
+            transaction_list = self.transaction_dao.get_all_transactions(account_id)
+            if transaction_list is None:
+                logging.warning("SAL method get all transactions, no transactions found")
+                raise FailedTransaction("No transactions found, please try again!")
+            else:
+                logging.info("Finishing SAL method get all transactions")
+                return transaction_list
 
     def service_delete_transaction(self, transaction_id: int) -> bool:
         logging.info("Beginning SAL method delete transaction")
-        result = self.transaction_dao.delete_transaction(transaction_id)
-        logging.info("Finishing SAL method delete transaction")
-        return result
+        if type(transaction_id) != int:
+            logging.warning("SAL method delete transaction, transaction ID not an integer")
+            raise FailedTransaction("The transaction ID field must be an integer, please try again!")
+        else:
+            valid_transaction_check = self.transaction_dao.get_transaction_by_id(transaction_id)
+            if valid_transaction_check is not None:
+                result = self.transaction_dao.delete_transaction(transaction_id)
+                logging.info("Finishing SAL method delete transaction")
+                return result
